- Change type of UserType: Student, Teacher ,… to long

+++ eVoterShare/src/evoter/share/model/UserType.java
@@ -19,9 +19,9 @@ public class UserType implements Serializable{
 	 */
 	private static final long serialVersionUID = -5893911091454051698L;
 	
-	public static final int SECRETARY =1;
-	public static final int TEACHER =2;
-	public static final int STUDENT =3;
+	public static final long SECRETARY =1;
+	public static final long TEACHER =2;
+	public static final long STUDENT =3;
 	private long id;
 	private String userTypeValue;
 	public long getId() {
